Version 2.3.1
Added the function OpenFile() to open referenced file (mapped to &#039gf&#039)
Use \"gf\" to open the file under the cursor. Graphic files are opened in \"feh\", pdf files in \"zathura\" and MS/OOO docs in \"LibreOffice\". Other filetypes are opened in VIM for editing. All this can be changed by editing the function OpenFile() in the file \"hyperlist.vim\".